Backend/BackendPython/requirements.txt
Django>=5.2,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
PyJWT>=2.8,<3.0
pika>=1.3,<2.0
psycopg2-binary>=2.9,<3.0
